+++ primal-count/src/prime_count.rs
@@ -1,7 +1,4 @@
-use crate::util::integer_square_root;
-use crate::util::integer_cubic_root;
-use crate::util::integer_quartic_root;
-use std::io;
+use crate::util::{integer_square_root, integer_cubic_root, integer_quartic_root};
 use std::collections::HashMap;
 use std::cmp::Ordering::*;
 
@@ -31,15 +28,6 @@ fn create_prime_array(bound: usize) -> Vec<usize> {
     return vec;
 }
 
-fn meissel_fn_small(m: usize, n: usize, prime_array: &Vec<usize>) -> usize {
-    // The number of numbers <= m that are coprime to the first n prime numbers.
-    if n == 0 {
-        return m;
-    } else {
-        return meissel_fn_small(m, n-1, &prime_array) - meissel_fn_small(m / prime_array[n-1], n-1, &prime_array);
-    }
-}
-
 // def _meissel_function_large(self, m, n):
 //     """The number of numbers <= m that are coprime to the first n prime numbers.
 //     Run for larger values where repeating isn't going to happen often. Use for n > 1000 or so"""
@@ -66,53 +54,23 @@ fn meissel_fn_small(m: usize, n: usize, prime_array: &Vec<usize>) -> usize {
 //         for M in stacks[0]:
 //             result += M*stacks[0][M]
 //     return result
-fn meissel_fn_large(m: usize, n: usize, prime_array: &Vec<usize>) -> usize {
-    meissel_fn_small(m, n, &prime_array)
+fn meissel_fn_large(m: usize, n: usize, prime_array: &Vec<usize>, meissel_cache: &mut HashMap<(usize, usize), usize>) -> usize {
+    meissel_fn_small(m, n, &prime_array, meissel_cache)
 }
 
-fn meissel_memoized(m: usize, n: usize, prime_array: &Vec<usize>, meissel_cache: &mut HashMap<(usize, usize), usize>) -> usize {
+fn meissel_fn_small(m: usize, n: usize, prime_array: &Vec<usize>, meissel_cache: &mut HashMap<(usize, usize), usize>) -> usize {
     if n == 0 {
         return m;
     }
     match meissel_cache.get(&(m, n)).map(|entry| entry.clone()){
         Some(result) => result,
         None => {
-            let value = meissel_memoized(m, n-1, &prime_array, meissel_cache) 
-                        - meissel_memoized(m / prime_array[n-1], n-1, &prime_array, meissel_cache);
+            let value = meissel_fn_small(m, n-1, &prime_array, meissel_cache) 
+                        - meissel_fn_small(m / prime_array[n-1], n-1, &prime_array, meissel_cache);
             meissel_cache.insert((m, n), value);
             return value;
         }
     }
-}
-
-fn num_primes_less_than(bound: usize) -> isize {
-    // Initialise prime array:
-    if bound < 1000 {
-        return create_prime_array(bound).len() as isize;
-    }
-    let sqrt_bound = integer_square_root(bound);
-    let primes = create_prime_array(sqrt_bound);
-
-    let a = num_primes_less_than(integer_quartic_root(bound));
-    let c = num_primes_less_than(integer_cubic_root(bound));
-    let b = num_primes_less_than(sqrt_bound);
-    println!("a,b,c={},{},{}", a,b,c);
-    let mut result = meissel_fn_small(bound, a as usize, &primes) as isize + ((b + a - 2) * (b - a + 1)) / 2;
-    println!("result={}", result);
-
-    for i in a..b {
-        let ith_prime = primes[i as usize];
-        result -= num_primes_less_than(bound / ith_prime);
-        if i < c {
-            let bi = num_primes_less_than(integer_square_root(bound / ith_prime));
-            for j in i..bi {
-                let jth_prime = primes[j as usize];
-                result += j - num_primes_less_than(bound / ith_prime / jth_prime);
-            }
-        }
-    }
-
-    return result
 }
 
 fn num_primes_less_than_memoized(bound: usize, primes: &Vec<usize>, prime_cache: &mut HashMap<usize, usize>, meissel_cache: &mut HashMap<(usize, usize), usize>) -> usize {
@@ -140,7 +98,7 @@ fn num_primes_less_than_memoized(bound: usize, primes: &Vec<usize>, prime_cache:
             let b = num_primes_less_than_memoized(sqrt_bound, primes, prime_cache, meissel_cache);
 
             // Issues with underflow here if b + a < 2
-            let mut result = meissel_memoized(bound, a, &primes, meissel_cache) + ((b + a - 2) * (b - a + 1)) / 2;
+            let mut result = meissel_fn_small(bound, a, &primes, meissel_cache) + ((b + a - 2) * (b - a + 1)) / 2;
 
             for i in a..b {
                 let ith_prime = primes[i];
@@ -162,7 +120,7 @@ fn num_primes_less_than_memoized(bound: usize, primes: &Vec<usize>, prime_cache:
 }
 
 // Top level function
-fn num_primes_less_than_main(bound: usize) -> usize {
+pub fn num_primes_less_than_main(bound: usize) -> usize {
     let primes = create_prime_array(integer_square_root(bound));
     let mut value_cache = HashMap::new();
 
@@ -181,103 +139,4 @@ fn num_primes_less_than_main(bound: usize) -> usize {
     let value = num_primes_less_than_memoized(bound, &primes, &mut value_cache, &mut meissel_cache);
     println!("Value = {}", value_cache[&bound]);
     return value;
-}
-
-
-
-fn main() {
-    // println!("How many primes below: ");
-    // let mut input = String::new();
-
-    // io::stdin().read_line(&mut input)
-    //     .expect("Failed to read line");
-    // let input: usize = input.trim().parse()
-    //     .expect("Couldn't turn value into an integer...");
-
-    // println!("There are {} primes below {}", create_prime_array(input).len(), input);
-    // println!("There are {} primes below {}", num_primes_less_than_main(input), input);
-    println!("There are {} primes", num_primes_less_than_main(10_000_000_000));
-
-
-}
-
-#[test]
-fn test_int_sqrt() {
-    assert_eq!(integer_square_root(0), 0);
-    assert_eq!(integer_square_root(1), 1);
-    assert_eq!(integer_square_root(16), 4);
-    assert_eq!(integer_square_root(17), 4);
-    assert_eq!(integer_square_root(24), 4);
-    assert_eq!(integer_square_root(587 * 587 - 1), 586);
-}
-
-#[test]
-fn test_int_cbrt() {
-    assert_eq!(integer_cubic_root(0), 0);
-    assert_eq!(integer_cubic_root(1), 1);
-    assert_eq!(integer_cubic_root(26), 2);
-    assert_eq!(integer_cubic_root(27), 3);
-    assert_eq!(integer_cubic_root(28), 3);
-    assert_eq!(integer_cubic_root(587 * 587 * 587 - 1), 586);
-}
-
-#[test]
-fn test_int_quartic_root() {
-    assert_eq!(integer_quartic_root(0), 0);
-    assert_eq!(integer_quartic_root(1), 1);
-    assert_eq!(integer_quartic_root(15), 1);
-    assert_eq!(integer_quartic_root(16), 2);
-    assert_eq!(integer_quartic_root(17), 2);
-    assert_eq!(integer_quartic_root(587 * 587 * 587 * 587 - 1), 586);
-    assert_eq!(integer_quartic_root(587 * 587 * 587 * 587 + 1), 587);
-}
-
-#[test]
-fn test_prime_array() {
-    let prime_array = vec![2, 3, 5, 7, 11, 13, 17, 19];
-    assert_eq!(create_prime_array(19), prime_array);
-    assert_eq!(create_prime_array(20), prime_array);
-}
-
-#[test]
-fn test_meissel_fn_small() {
-    let prime_array = vec![2, 3, 5, 7, 11, 13, 17, 19];
-    assert_eq!(meissel_fn_small(30, 8, &prime_array), 3);
-    assert_eq!(meissel_fn_small(100, 1, &prime_array), 50);
-}
-
-#[test]
-fn test_meissel_fn_large() {
-    let prime_array = vec![2, 3, 5, 7, 11, 13, 17, 19];
-    assert_eq!(meissel_fn_large(30, 8, &prime_array), 3);
-    assert_eq!(meissel_fn_large(100, 1, &prime_array), 50);
-}
-
-
-#[test]
-fn test_meissel_fn_memoized() {
-    let prime_array = vec![2, 3, 5, 7, 11, 13, 17, 19];
-    let mut cache = HashMap::new();
-    assert_eq!(meissel_memoized(30, 8, &prime_array, &mut cache), 3);
-    assert_eq!(meissel_memoized(100, 1, &prime_array, &mut cache), 50);
-}
-
-#[test]
-fn test_num_primes_less_than() {
-    assert_eq!(num_primes_less_than(7), 4);
-    assert_eq!(num_primes_less_than(100), 25);
-    assert_eq!(num_primes_less_than(2143), 324);
-    assert_eq!(num_primes_less_than(1_000_000), 78_498);
-    // assert_eq!(num_primes_less_than(1_000_000_000), 50_847_534);
-}
-
-#[test]
-fn test_num_primes_less_than_main() {
-    assert_eq!(num_primes_less_than_main(7), 4);
-    assert_eq!(num_primes_less_than_main(100), 25);
-    assert_eq!(num_primes_less_than_main(2143), 324);
-    assert_eq!(num_primes_less_than_main(1_000_000), 78_498);
-    assert_eq!(num_primes_less_than_main(1_000_000_000), 50_847_534);
-    // assert_eq!(num_primes_less_than_main(1_000_000_000_000), 37_607_912_018);
-    // assert_eq!(num_primes_less_than_main(1_000_000_000_000_000), 29_844_570_422_669);
 }
