+++ primal-count/src/util.rs
@@ -1,4 +1,4 @@
-fn integer_square_root(value: usize) -> usize {
+pub fn integer_square_root(value: usize) -> usize {
     // Returns the largest integer at least sqrt(n) using Newton's method
     let mut x = value;
     let mut y = (x + 1) >> 1;
@@ -9,7 +9,7 @@ fn integer_square_root(value: usize) -> usize {
     return x;
 }
 
-fn integer_cubic_root(value: usize) -> usize {
+pub fn integer_cubic_root(value: usize) -> usize {
     // Returns the largest integer at least cbrt(n) using Newton's method
     let mut x = value;
     let mut y = (2 * value + 1) / 3; // Plus 1 only to protect against division by 0 later - not because it's right
@@ -20,7 +20,7 @@ fn integer_cubic_root(value: usize) -> usize {
     return x;
 }
 
-fn integer_quartic_root(value: usize) -> usize {
+pub fn integer_quartic_root(value: usize) -> usize {
     // Returns the largest integer at least n^(1/4) using our fast integer sqrt
     return integer_square_root(integer_square_root(value));
 }
